+++ lib/view/bank_button_area_view.dart
@@ -1,4 +1,8 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
+
+// Project imports:
+import 'package:simple_sample_manager/constant/bank_button.dart';
 
 class BankButtonArea extends StatelessWidget {
   const BankButtonArea({Key? key}) : super(key: key);
@@ -29,35 +33,6 @@ class BankButtonArea extends StatelessWidget {
             );
           }).toList(),
         ),
-      ),
-    );
-  }
-}
-
-class BankButton extends StatelessWidget {
-  final String _keyWord;
-  const BankButton(
-    this._keyWord, {
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      style: ElevatedButton.styleFrom(
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(0), //角の丸み
-        ),
-        foregroundColor: Color.fromARGB(255, 0, 195, 255),
-        backgroundColor: Color.fromARGB(127, 0, 195, 255),
-        fixedSize: Size.fromHeight(20),
-      ),
-      onPressed: () {
-        //ボタンが押されたら「textfield.dart」のDisplauControllerに「_keyWord」を追加する。
-      },
-      child: Text(
-        _keyWord,
-        style: const TextStyle(fontSize: 20.0),
       ),
     );
   }
