+++ lib/view/footer_area_view.dart
@@ -1,4 +1,7 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
+
+// Project imports:
 import 'package:simple_sample_manager/constant/menu_button.dart';
 
 class FooterArea extends StatelessWidget {
