+++ lib/view/pad_button_area_view.dart
@@ -1,5 +1,9 @@
+// Flutter imports:
 import 'package:flutter/material.dart';
+
+// Project imports:
 import 'package:simple_sample_manager/constant/edit_button.dart';
+import 'package:simple_sample_manager/constant/pad_button.dart';
 
 class PadButtonArea extends StatelessWidget {
   const PadButtonArea({Key? key}) : super(key: key);
@@ -44,35 +48,6 @@ class PadButtonArea extends StatelessWidget {
           ],
         ),
       ],
-    );
-  }
-}
-
-class PadButton extends StatelessWidget {
-  final String _keyWord;
-  const PadButton(
-    this._keyWord, {
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      style: ElevatedButton.styleFrom(
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(0), //角の丸み
-        ),
-        foregroundColor: Color.fromARGB(255, 0, 195, 255),
-        backgroundColor: Color.fromARGB(127, 0, 195, 255),
-        fixedSize: Size.fromHeight(20),
-      ),
-      onPressed: () {
-        //ボタンが押されたらメインディスプレイにサンプル情報を表示させる。
-      },
-      child: Text(
-        _keyWord,
-        style: const TextStyle(fontSize: 20.0),
-      ),
     );
   }
 }
